Improved error message in case of missing 'GenerateReactions:' from input file
We now check for a GenerateReactions: yes/no line in the input file
and tell the user to add one if it's missing.

This assumes that everything else up until that point in the input file
worked perfectly and was error free, otherwise this message may well
be reported when in fact something else was wrong.

This should be a slight improvement, at least.