Updated PopulateReactions
In the RMG-meeting on 25Feb2010, Bill and others mentioned that the
PopulateReactions module should also be able to read in kinetics
from a PrimaryReactionLibrary (soon to be named PrimaryKineticsLibrary).

Before: the RMG_database would be read immediately after reading in the
system temperature.
Now: the RMG_database is read immediately before the module reacts all
species given in the input file.

Next step: Have module read in PrimaryReactionLibrary (as specified by
user) before reading in RMG_database