Added daspk private submodule for code we can't distribute
The contents of /source/daspk/private is required to build daspk,
but we do not have the right to distribute it. It is thus kept in a
separate git repository, stored on the green group's private server.
Only people who can ssh to this server will be able to check out the files.
The master git repository (superproject) which we can share with the world
will contain a reference to which commit from this submodule is required.

To fiddle with this or edit these files, these could be handy:
Submodule manual
http://www.kernel.org/pub/software/scm/git/docs/git-submodule.html
Useful tutorial
http://git.wiki.kernel.org/index.php/GitSubmoduleTutorial