Added user option to display SMILES string
In the condition.txt file, the user now has the option to display a
blank SMILES string in the thermochemistry portion of the chem.inp
file (for each species).  The user enters this information in the
optional "SMILES" field within the "ChemkinUnits" portion of the input
file; the options are "off" or "on".  The default is off.