InChI debugging (11MAR2010 1527)
Removed reading the standard output of the InChI process.  On Linux, this
block is apparently never reached.  The reason for the concern is that
"error" messages are printed to the standard output files from submitting
a job on Linux.  Technically, there is no problem but this could worry
RMG users.