When GUI opens existing condition file, sets the full database path.
Although this makes it look to the user like the absolute path has been specified
in the file (whereas in fact it's specified relative to $RMG/) it has the benefit
of telling the GUI where to look when selecting additional primary libraries 
and seed mechanisms. This is an unfortunate trade-off and I'm open to argument.