More informative error message for incorrect databases
When we declare a node is a child of another node (in a tree)
the code actually checks to see if this is true. When it is
not a child, it prints a warning message to the screen.
Since 2005 this has been "Father __ Child __", which means
so little to anyone that we didn't realize it was even a warning.

Now maybe someone will be motivated to fix the dictionaries!

Richard West rwest@mit.edu