Allowing RMG to handle multiple kinetics: Debug1
MRH is working on changing the way RMG stores kinetics for a reaction.
Currently, we have a "Kinetics" and "additionalKinetics" field.  Two
problems that arise are:

(1) What if more than 2 kinetics are supplied by the user (e.g. in a
    Seed Mechanism) for a given reaction.  RMG would print a message
    stating only 2 can be handled and then terminate the program.
(2) If supplying duplicate kinetics to RMG through the Primary Reaction
    Library, RMG will only recognize the first (as there is no method
    readInAndStoreAdditionalKinetics)

MRH is still working out the bugs, in particular reading/writing the
Restart files.