Better handling of negative activation energies in FAME.
FAME used to simply set negative values of Ea to zero when they were
encountered. This is usually okay because negative Ea values usually are low
in magnitude. However, they can occasionally be high in magnitude due to
poor database values. In this case, it is better to modify A by a factor of
exp(-Ea/RT) *and* set Ea to zero, so that the value of k(T) at that T is
unchanged.

The treatment of negative Ea is sufficient to allow for removal of the
associated warning message. RMG still warns about negative Ea, but only at
the kinetics estimation step, not when evaluating p-dep networks.