Updated condition file for more flexible pressure dependent options.
You can now specify a TRange and PRange and let RMG distribute the actual
temperatures and pressures. For Chebyshev polynomial interpolation models,
the Gauss-Chebyshev grid is used. For other models, the points are evenly
distributed on 1/T and log P scales.

You can instead specify a set of Temperatures and Pressures manually. You
should specify one each of TRange or Temperatures and PRange or Pressures.
Strictly speaking, this is not required, as RMG will give a default set of
temperatures and/or pressures if not specified.

You will probably want to use TRange and PRange unless you have a specific
reason to specify the temperatures and/or pressures manually.

Also updated the documentation to reflect the changes to the condition file
format.