Now using specified interpolation model to evaluate k(T,P) within RMG.
Before we would always use bilinear interpolation within RMG, and simply
provide the interpolation model as output (e.g. to CHEMKIN). However, it
turns out that the grid of temperatures and pressures varies with the
interpolation model. In particular, the optimum grid for Chebyshev polynomial
models does not place points at the extrema, even though the fit is still
valid there.

Thus, we simply use the desired interpolation model everywhere. This makes
logical sense as well, as we want to be using the same k(T,P) function in
everything we use it for.