Adding comments to pdep kinetics for reverse rxns
When reporting pdep kinetics in the chem.inp file, a comment was
written for where the high-p-limit kinetics came from (if applicable).
For pdep reactions whose high-p-limit kinetics came from a reverse
reaction + equilibrium, the comment was not being stored
w/the reaction (MRH may have accidentally erased RW work when merging the
dupKinetics branch).  In any case, this issue is now resolved.