Bug fix regarding multiple kinetics & pdep
When reporting comments for a pdep reaction, RW noticed that some
comments were not preceded with an exclamation mark.  MRH believes
this is only the case when multiple kinetics contribute to the overall
reaction rate (e.g. 1-butyl --> 2-butyl via a 3- and 4-membered TS).
The issue has been resolved.

This commit closes Issue 36.

Furthermore, more detailed comments are provided in the case of a
pdep reaction whose rate came from a summation of multiple kinetics.