New input file (pdep) option: DecreaseGrainSize
After a fame calculation, RMG checks whether any of the pdep rates
exceed the high-p-limit rate by more than a factor of 2.  If so, the
number of grains is increased (grain size is decreased) and the
pdepnetwork is re-run in fame.

JWA suggested to have this feature be an option for the user (since
this increases the runtime of a RMG simulation drastically).  MRH has
added a new field "DecreaseGrainSize" to the input file (for both RMG
and PopulateReactions module).  The default value is off.