In fame, fixed dynamic memory allocation error during determineEnergyGrains().
The density of states and equilibrium distribution arrays in the isomer being
used to determine the energy grains were not getting deallocated and
reallocated to the larger size when the search takes multiple iterations.
This caused an infinite loop. Code has been added to properly handle the
memory reallocation and to prevent the infinite loop from occurring in the
future.