Adding forbiddenGroups.txt file to Diels_alder_addition
RMG threw a RateConstantNotFoundException when reacting a fused ring
against isobutylene (see Issue #38).

This commit resolves the SPECIFIC issue that Yushi was running into by
adding a forbiddenGroups.txt file to the Diels_alder_addition reaction
family directory.  The two forbidden structures do not allow the *3 atom
to be bonded to the *5 atom, nor the *6 atom to be bonded to the *4 atom.