Added check for null pointer exception when enlarging a PDepNetwork.
If a network is identified as the object that needs enlarging, it is supposed
to determine the nonincluded isomer with the highest leak flux and generate
all reactions involving that isomer. However, the code as previously written
could return null for the isomer to explore, which would cause an unhelpful
NullPointerException to be thrown in Java. This commit checks for this
exceptional case and prints a more verbose error message if found.