RMG now reports species of interest's conversion
Before: At the end of each ODE time step, RMG would report the conversion
of the first species listed in the condition file.  However, the first
species did not always correspond to the "species of interest," namely,
the species listed in the "Goal Conversion" field.

Now: RMG will report the conversion of the species present in the "Goal
Conversion" field, regardless of the location in the condition file.  If
the goal of the simulation is "ReactionTime," RMG reports the conversion
of the first species listed in the condition file (as before).