Fixed handling of resonance structures for transport properties
In the event of a species containing resonance structures, RMG should
check whether any of the chemgraphs resonance structures matches one
in a Primary Transport Library.

Before: RMG would check the first structure and return TransportData.
MRH had the code exit the loop if TransportData was not null.  However,
it should always return not null (because the transport data will be
estimated in the event it is not found in a primary transport data).

Now: RMG will check whether the returned transport data is
"fromprimarytransportlibrary".  If so, this is the data we want for
the species and it will be saved.  If not, the loop over the resonance
structures continues.  If none of the resonance structures are found
in a primary transport library, the transportdata for the "most stable"
chemgraph (as selected when estimating the transport properties) will
be employed.