Corrected formatting for reporting data in tran.dat
According to the CHEMKIN-v4.1.1 Manual, the syntax of the species'
transport file is free-form (other than the species name being limited
to 16 characters).  So, the previous formatting should be correct,
technically.

MRH is updating the syntax to the tran.dat format he has been using
in his butanol study, which has been tested against CHEMKIN-vII,
CHEMKIN_v4.1.1, CHEMKIN-Pro, and CHEMKIN-MFC.