Updating the format of species names in Restart files
Before: When writing species' names to the restart files, the
toChemkinString() method was used.  This invariably wrote many of the
species as SPC(#).  Thus, when reading in the restart files, RMG would
print "Making species: SPC(#)" or "New reaction: SPC(#)+SPC(#)=SPC(#)"

Now: When writing species' names to the restart file, the
toRestartString() is used.  The name syntax for this method is to
write the chemical formula + the ID (inside of parenthesis).  Thus,
RMG's printing to screen when reading in Restart files displays more
information.

Closes #14