Bug fix for PopulateReactions module: Reporting same rxn twice
Suppose a user inputs CH3, H2, H, and CH4 in the input file for the
PopulateReactions module.  The module will find both CH3+H2=H+CH4
and H+CH4=CH3+H2.  We only want to report this reaction (and its
kinetics) once in the output file.

Before: RMG would check each non-pdep reaction against all other non-pdep
reactions to ensure only one instance of the reaction is returned.  It was
checking for duplicates by comparing:

 - the forward structure of the reaction in question against the forward
   structure of all other non-duplicated non-pdep reactions
 - the reverse structure of the reaction in question against the reverse
   structure of all other non-duplicated non-pdep reactions

Obviously, if it fails the first, it will fail the second, deeming the
reaction as not a duplicate.

Now: The second criteria has been changed to compare:
 - the reverse structure of the reaction in question against the forward
   structure of all other non-duplicated non-pdep reactions