Bug fix: Checking whether PrimaryTransportLibrary is null
When generating transport properties for a species, the chemgraph is first
compared against all primary transport libraries.  However, there was no
check in the code to ensure a primary transport library exists.  In
particular, when running the ThermoDataEstimator module (which does not, nor
should not, require the locations of any primary transport libraries), an
error would terminate the program.

The code now checks whether a primary transport library exists before
attempting to compare a chemgraph against the library.