Ensuring all species/reactions are either in the core or edge
Background: When running a RMG simulation w/the ReadRestart flag on, an
error may have been thrown relating to RMG's failure to properly read
the pdepnetworks.txt file (see Issue 8).

MRH determined that for all occurrences (RW, JDMO, Shamel), the problem was
that some of the species writeen to the pdepnetworks.txt file were not
stored in either the coreSpecies.txt or edgeSpecies.txt files.  It was
further determined (through discussions w/JWA + RMG developers) that there
were occurrences in a RMG simulation where a species would not be
labeled as either "core" or "edge" (a similar situation arose for reactions).

This rare event occurred when the leak flux of a network was > the model's
characteristic flux, in addition to no edge species' flux being > the
model's characteristic flux.

Now, when an isomer is "explored" within a PDepNetwork and creates new
reactions, that reaction is classified as either "core" or "edge".  Any
new species generated (and not already present in the model) are labeled
as "edge".

A new elseif statement was added to CoreEdgeReactionModel's
addUnreactedReaction() method.  In the event of the rare occurrence
mentioned above, an "edge" species is making an "edge" species, a process
which would never occur otherwise in a RMG simulation.