extended AbramData to also include V parameter (now S,B,E,L,A,V)
Previously it was just S,B,E,L,A (although one constructor had V in place of L)
The AbrahamGAValue class which this extends, still only has the first five.
That is because V is McGowan's Volume, which is not determined by the group
additive scheme of Platts et al. that is used for the others.