Remove code duplication in writeChemkinInputFile methods
Most of this method was duplicated with some small differences.
Now one method (called with a ReactionSystem) just calls the
other method (called with a ReactionModel and SystemSnapshot).

The underlying writeChemkinPdepReactions(ReactionSystem) that
was called by the removed code is now deprecated, and has been removed.
This code is slightly different from the alternative
writeChemkinPdepReactions(Reactionmodel, SystemSnapshot) which is now
always used.

Mike and I think the removed code is less complete than the code we left in its place.
For example, the removed code didn't report the units for activation energy, although
unit conversion was taking place underneath (THIS WAS PROBABLY A BUG)

Previously, we think one version was used at intermediate simulations, and
the other was used at job completion. Now there is only one to maintain.