Save intermediate Chemkin files in numbered subfolders.
WARNING!: At the start of each run the [workingDirectory]/chemkin/ directory
is now wiped clean. This happens to other such directories so shouldn't be too
unexpected, but is a change, so be warned.

The chemkin, transport, and table of rate coefficients, are all written to
numbered subfolders in each simulation step (numered according to number of core species)
and also copied into the chemkin folder (so that the chemkin folder always contains
the latest version).