Simplify code in writeChemkinPdepReactions
After scratching my head for a while I determined that this section of code 
has evolved over many different generations and commits, to be 
a long complicated way of doing very little. I have simplified it a bit.

The history was to do with detecting duplicates, but a pdep and a non-pdep
do not count as duplicates, according to chemkin (as described by Mike in 
88a467476c5 and fa5ac7f1bb9 )