Fixed misunderstanding in pressure dependent network enlarging procedure.
The key is to treat the nonincluded unimolecular isomers as infinite sinks.
This ensures that the rate constants for reactions to the nonincluded isomers
are overpredictions because they include all possible results of formation of
an excited molecule of that isomer. Fortunately much of this functionality
already existed within FAME; the main effect of this commit is to change
slightly the information passed to FAME so it calculates things properly.

As a bonus, this should cause FAME calculations to go faster, because the
master equation matrix will be significantly smaller for networks containing
multiple isomers. Furthermore, the RS method should be less likely to be
unsuccessful for the same reason.