Possible fix for NullPointerException encountered when enlarging PDepNetworks.
Based on the extra output generated as a result of changes made in the last
commit, the method of enlarging the model core and the partial networks is
not quite robust as a result of the recent changes to handling nonincluded
isomers. This short change may address this; it certainly does not hurt
anything.