PDep: better merging of networks at network enlarge step.
When a PDepNetwork leak flux is the highest flux, an isomer in that network
is chosen to be explored. If that isomer also exists as an isomer in any other
partial network, that network should be merged into the one being enlarged.
This is now done explicitly before exploring the reactions of the isomer.

This is a better implementation of the last PDep commit.