BUG FIX: Primary Reaction/Kinetics Library & resonance structures
When RMG checks each new reaction against the Primary Reaction (soon to
be Kinetics) library, the algorithm now iterates over all possible
resonance structures of the given chemgraph (if applicable).  This was the
suggested solution of Greg and Richard.

I've tested this against the example Greg mentioned in the Issue tracker
and the problem has been resolved.

Closes #48