Adding additional comments: Reactions whose rates come from Primary Library
When the structure of a reaction matches one present in a Primary Reaction
(soon to be Kinetics) Library, the user-specified kinetics are used rather
than the RMG estimate.  To signify this in the chemkin file, the "comments"
would read "Primary Reaction Library: ____"

Now, the comments have an additional piece of information, namely the
(forward) reaction family template that RMG would have matched, e.g.

"R_Recombintion: Primary Reaction Library: ____"