Updated field tag names/placement in RMG GUI
Some of the tab names have been changed to more accurately reflect what
information the user is expected to enter.  Furthermore, some of the
fields have moved; these moves are in hopes of grouping fields in a more
logical sense (i.e. dassl/daspk information is not grouped with pruning
options).

Closes #50