BUG FIX: Update Reaction Lists when species is added to core
When running a pdep job, the core can either be increased by adding a species
from the edge to the core OR by exploring a network (which has the largest
leak flux).

For one example (as mentioned in the Issue tracker), RMG decided that the
methane (CH4) pdep network had the greatest leak flux; however, both sets
of isomers, CH4 and CH3+H, were already included, meaning there should be
no NonIncludedRxn in that particular network (and hence, no leak flux).

JWA and MRH determined that when a species was added from the edge to the
core, RMG was not "refreshing" the pdep networks.  RMG now refreshes the
networks, via a call to the network.updateReactionLists method, after
adding a species from the edge to the core AND after exploring the network
with the greatest leak flux.

Closes #60