Adding additional constructors for PDepIsomer class
MRH is working on reading/writing Restart files with the new format (i.e.
NonIncludedReactions can be irreversible, additional comments in the
Isomer.toString() method, etc.)

When reading the Restart files and making Isomers, the "included" boolean
for each Isomer was assumed, as defined in the PDepIsomer class (i.e.
included for Isomers w >1 reactant and not included for Isomers w =1
reactant).  Now, this information is written to the pdepnetworks.txt file
and read in by RMG.  Thus, to preserve the status of each Isomer, these
new constructors have been written.