BUG FIX: Cleaning up reading of Restart/pdepnetworks.txt file
MRH has replaced all sd.getSpeciesFromChemkinName() calls to
sd.getSpeciesFromNameID().

The .getSpeciesFromChemkinName() call could lead to a bug: Suppose we are
reading in species C2H5O2J(67).  This name has 11 characters so RMG would
replace the ChemkinName w/ SPC(67).  If RMG then compares the name of the
species (as written in all Restart files, which is C2H5O2J(67)) with the
ChemkinName, it will not find it and throw an error.  The
.getSpeciesFromNameID() method looks for the chemical formula (name) + ID.