Adding getReactionLibrary() method to LibraryReactionGenerator class
In one instance in RateBasedPDepRME.java, there is a check to see if
the LibraryReactionGenerator is null.  However, MRH believes this will
never be null (even if a user does not supply a ReactionLibrary).

However, what will be null is the reactionLibrary variable of the
LibraryReactionGenerator class.  Now RMG checks whether the reactionLibrary
is null or not.