BUG FIX: Adding explored species to the CoreEdgeReactionModel edge
History lesson: When writing the Restart/pdepnetworks.txt file, there were
some instances when a species present in a pdepnetwork would not be
written to either the coreSpecies.txt or edgeSpecies.txt files.  To account
for this, MRH added code to PDepNetwork.java to check whether every new
explored species should be added to the edge.

Problem: MRH used the .addReaction() method to accomplish this.  While all
species were now being specified as either core or edge, this also added
unimolecular reactions (e.g. A --> B or A --> C + D) to the edge of the
CoreEdgeReactionModel.  Thus, RMG would have incorrect duplicate instances
of these reactions: the pdep reactions in a network (correct) and the high-P
limit reactions in the CoreEdgeReactionModel (incorrect).  RMG would then
double-count the flux of each of these reactions and report both in the
chem.inp file (thanks to Shamel for finding this).

Resolution: All newly explored species are still added to the edge, but the
newly explored reactions are left alone (and remain in the pdepnetwork).

Closes #61