Much improved method for counting core and edge reactions and species.
I've collected the various versions of the printing of the model size into a
single function ReactionModelGenerator.printModelSize(). The counting is a
bit involved; see that function for details.

Also added functions for counting the edge reactions in the PDepNetworks. Only
reactions with both reactant and product isomers as included are eligible.