Improving how RMG handles duplicate kinetics from multiple rxn templates
This commit relates to Issue #65 (but does not close it).  When RMG calls
the TemplateReactionGenerator.react() method, it adds reactions to the
whole list of reactions one reaction family at a time.  Thus, if a reaction
(i.e. A + B --> C) can occur via multiple reaction families, only the first
will be saved while the others will be ignored.  How we want RMG to handle
this is to recognize the reaction structure already exists and to add
additional kinetics to this reaction.  Something similar already exists in
the SeedMechanism class.

MRH has taken the bits of code in SeedMechanism that pertain to identifying
dupliate reactions / placing kinetics in the proper place and made a new
method (so that the TemplateReacitonGenerator class can call this).  This is
step one towards resolving Issue #65.