Cleaning up duplicate code in TemplateReactionGenerator class
The TemplateReactionGenerator class has two different .react() methods:
one if a LinkedHashSet is passed (i.e. all initial species in the core) and
another if a LinkedHashSet and a Species is passed (i.e. every time one new
speices is added to the core).

The former has been restructured to call the latter.  Moving closer towards
resolving Issue #65.