Added batch scripts for running RMG and PopulateReactions to examples.
These are best used by copying them to the directory containing the
condition file and double-clicking them to run. They will perform a few
checks for appropriate executables and input files, run the job, and
save the results to a log file. There is not much feedback on the
console unless there is a missing executable or input file. Also, a
failed job will look like a successful job, except that the log file
will indicate that the job had some error.