adjusted parameters for pruning test case
in example run, it didn't seem to be pruning, at least not early on, so I increased TerminationTolererance and reduced Goal Conversion
completed in a little over 275 minutes on Pharos with:
Memory used: 1550712832
Free memory: 893657096
The model core has 2519 reactions and 166 species.
The model edge has 67663 reactions and 12091 species.
Model Generation Completed