(Temporary fix) RateConstantNotFoundException
RMG will fail with a "RateConstantNotFoundException" when it cannot
estimate a forward, nor reverse, reaction rate.  The reason for this is not
entirely understood (see Issue #38 for more details).

Rather than fail, MRH now sets the Arrhenius expression equal to 0.0 (as
suggested by RW).  MRH has also had to change some "if k > 0" or
"if rate > 0" to include zero.