Avoid expanding altered PDep networks with multiple reaction systems.
When running with multiple reaction systems, one reaction system may mark a 
species to be added to the core and another will mark a network to be 
expanded. Adding the species to the core could alter the network. When
the network is then expanded, an error occurs because it has already
*been* expanded (or at least altered).

We now check, before expanding a network, that it has not been altered
already in this step.

Closes #70.