Only report Library Reactions containing the newly added species.
When a species is added to the core, the LibraryReactionGenerator is 
used to react everything in the core. Previously it reported all the 
matching core reactions. Now it just lists the ones that contain the 
newly added species. 

This has been done in both RateBasedRME.java and RateBasedPDepRME.java.
The amount of code duplication and divergence is horrendous. I hate this.