PDep: Changes to bandedness of master equation collision matrices.
In mastereqn.f90, the subroutine collisionMatrix() will no longer do any
truncation. This means that we only have to choose the bandwidth in one
location.

The choice of bandwidth has been modified to choose the larger of dEdown and
R*T as the basis for the width, in order to incorporate a sufficient number of
grains at high temperatures. (This is for the RS method only, as MSC doesn't
use the collision matrix anyway.)