PDep: Fixed bug in Chebyshev polynomial fitting in FAME.
FAME was not using the correct minimum and maximum values in the range, as the
Gauss-Chebyshev grid points are placed such that the endpoints of Tlist and
Plist are *not* the same as the valid range (Tmin, Tmax) and (Pmin, Pmax). In
order to do this correctly, I modified RMG to pass the correct minimum and
maximum temperatures and pressures to FAME, and modified FAME to use these for
the Chebyshev polynomial fitting.

This was noticed due to a roundoff error observed when FAME tried to use the
minimum and maximum temperatures/pressures from the grid as the Chebyshev
endpoints, due to chebyshev polynomials being stricly undefined outside of the
range [-1, 1]. This commit avoids this issue because the grid endpoints are
sufficiently far from +1 and -1 so that roundoff can't hurt anything.

Closes #73.