Don't run FAME for networks that don't contain any included unimolecular isomers.
We only use the high pressure limit k(T) to determine the leak flux, and the
calculated nonincluded reaction doesn't really represent it because the isomer
is only a pseudospecies, so there's no point in running FAME for these. Almost
all networks start off this way, so the net effect of this is a nice speed
boost (and perhaps a slight memory savings too), with no change in the
generated model.