Modified calculation of equilibrium constant for solution phase chemistry.
The modified version checks if the solvation flag is being used and changes the
standard state concentration used to calculate Keq.

The numerical value of the std. state concentration to be used for is not yet
clear. It has been set to 1M for the moment based on literature values. This
value shall subsequently be updated.