Updated calculation of equilibrium constant in solution phase.
After analysis of the assumptions that go into the definition of deltaG_solv,
it seems that one need not correct for std states as they are assumed to have
the same magnitude in both the gas and the solution phase. Traditionally, this
concentration has been assumed to be 1 mol/L in both phases. However, at this
concentration, the assumption of an ideal gas phase is an extreme one. As a
result, it makes more sense to choose a more dilute std. state where the
assumptions of ideality in the gas phase and infinite dilution in the solvent
phase may be valid. The gas phase std. state concentration of 1/RT satisfies
this criterion better that the assumption of 1 mol/L. It should be noted that
in the calculation of deltaG_solv, one needs to care only about the ratio of
the standard states in the gas phase and the solution phase. However, in
calculations of the equilibrium constant, the solution phase std. state
concentration is used explicitly.