Added Methylformate reaction library (High-P limits; not seed mechanism)
This contains high-P limits for four reactions as provided by Franklin.
It should be useful as a Reaction Library. (At least in the current version
of RMG, reaction library reactions are sent through Fame for P-Dep calculation).
Don't use it as a seed mechanism, which assumes all pressure dependence is
already included.