BUG FIX: Handling duplicate species in multiple thermo_libraries
When a user supplies RMG with multiple thermo_libraries, there is a chance
the same species (e.g. H2) will appear more than once.  In these instances,
the thermo RMG should employ is the first thermo data read in.

However, RHW noticed that this was not the case, when supplying the RMG
default library and Franklin's DFT_QCI library.  The problem was that
H2, H, and O2 had identical names/graphs in both libraries and RMG did
not recognize these as duplicates.

RMG now recognizes this instance (same name, same graph, different thermo
library) and reports this to the user.  Also, the thermo from the first
read-in instance is implemented.

Closes #89